First of many patches to convert long's to int32_t for portability on 64 bit systems

+++ SF3000/SFlib/Bonus_Control.h
@@ -1,9 +1,10 @@
+#include <stdint.h>
 #include "Ship_Struct.h"
 
 extern ship_list ships;
 extern ship_stack *players_ship;
 
-extern long bonus_collision_this_frame;
+extern int32_t bonus_collision_this_frame;
 
 void bonus_control(ship_stack *);
