+++ SF3000/SFlib/Global_2.h
@@ -1,4 +1,6 @@
 
+#include <stdint.h>
+
 // Setup various structures
 #include "Ship_Struct.h"
 extern ship_list ships;
@@ -31,42 +33,42 @@ extern docking_struct docked;
 extern bonus_crystal_adder_struct bonus_crystal_adder;
 extern bonus_crystal_hill_struct bonus_crystal_hill;
 extern bonus_crystal_object_struct bonus_crystal_object;
-extern long bonus_collision_this_frame;
+extern int32_t bonus_collision_this_frame;
 
 // Players control rates
-extern long players_x_control;
-extern long players_y_control;
-extern long players_z_control;
-extern long players_thrust_control;
+extern int32_t players_x_control;
+extern int32_t players_y_control;
+extern int32_t players_z_control;
+extern int32_t players_thrust_control;
 
 // ########### bug fix for sound_conteol
-extern long laser_sound_counter;
+extern int32_t laser_sound_counter;
 
 // Two nearest ships of this type make engine sounds
 extern ship_stack *sound_big_ship;
 extern ship_stack *sound_small_ship;
-extern long sound_channel_small_ship;
-extern long sound_channel_big_ship;
-extern long sound_sample_small_ship;
-extern long sound_channel_shields_low;
-extern long sound_channel_beam_laser;
-extern long engine_sounds_on_or_wot;
+extern int32_t sound_channel_small_ship;
+extern int32_t sound_channel_big_ship;
+extern int32_t sound_sample_small_ship;
+extern int32_t sound_channel_shields_low;
+extern int32_t sound_channel_beam_laser;
+extern int32_t engine_sounds_on_or_wot;
 
 extern ship_stack *ship_being_viewed;
 
 // Rate of fire for ground objects - chances in 1024
-extern long ground_laser_rate;
-extern long ground_sam_rate;
-extern long ground_ship_rate;
+extern int32_t ground_laser_rate;
+extern int32_t ground_sam_rate;
+extern int32_t ground_ship_rate;
 
-extern long ground_laser_type;
-extern long ground_ship_type;
-extern long air_ship_type;
+extern int32_t ground_laser_type;
+extern int32_t ground_ship_type;
+extern int32_t air_ship_type;
 
 // Toggle for the ground lasers
-extern long laser_counter;
+extern int32_t laser_counter;
 
-extern long test_mode;
+extern int32_t test_mode;
 
 // Misc counters etc.
 extern long pod_counter;
@@ -77,8 +79,8 @@ extern long ship_viewed_last_frame_x_rot;
 extern ship_stack *ship_viewed_last_frame;
 
 // Setup data tables for quick reference look up
-extern long cosine_table[2048];
-extern long *sine_table;
+extern int32_t cosine_table[2048];
+extern int32_t *sine_table;
 extern long pex_table[16384];
 extern long pex_table_near[2048];
 extern long quick_height_table[256][4];
