+++ SF3000/SFlib/SF_CelUtils.h
@@ -5,6 +5,8 @@
 
 //	includes
 
+#include <stdint.h>
+
 #include "SF_Allocation.h"
 #include "displayutils.h"
 #include "graphics.h"
@@ -143,8 +145,8 @@ extern char *map32;
 extern char cache_lookup[64 * 64 * 2];
 extern char cache_free[CEL_MAXCREATION];
 extern ScreenContext *screen;
-extern long cosine_table[2048];
-extern long *sine_table;
+extern int32_t cosine_table[2048];
+extern int32_t *sine_table;
 
 // Function Prototypes
