+++ SF3000/SFlib/Update_Frame.h
@@ -1,3 +1,4 @@
+#include <stdint.h>
 
 #include "SF_CelUtils.h"
 #include "Ship_Control.h"
@@ -18,7 +19,7 @@ extern stick_struct pro_stick;
 /* External Definitions */
 
 // bugggggg fix
-extern long laser_sound_counter;
+extern int32_t laser_sound_counter;
 
 extern void rotate_node_from_c(void *);
 extern long wave_counter;
@@ -36,21 +37,21 @@ extern game_status status;
 extern long pod_counter;
 extern long *temp_store;
 
-extern long laser_counter;
+extern int32_t laser_counter;
 
 #include "SF_Mission.h"
 extern planet_data planet_info;
 
-extern long players_x_control;
-extern long players_y_control;
-extern long players_z_control;
-extern long players_thrust_control;
+extern int32_t players_x_control;
+extern int32_t players_y_control;
+extern int32_t players_z_control;
+extern int32_t players_thrust_control;
 
 extern long keypad_iomap[8];
 
-extern long bonus_collision_this_frame;
+extern int32_t bonus_collision_this_frame;
 
-extern long test_mode;
+extern int32_t test_mode;
 
 // Shitty get it ready for another show fix
 extern long plot_clouds_or_wot;
